+++ test/contracts/SportMarkets/SportsAMMDoubleChance.js
@@ -965,13 +965,6 @@ contract('SportsAMM DoubleChance', (accounts) => {
 			answer = await Thales.balanceOf(first);
 			console.log('Balance after buying: ', fromUnit(answer));
 
-			answer = await SportsAMM.canExerciseMaturedMarket(homeTeamNotLoseMarket.address);
-			console.log('Can exercise options homeTeamNotLoseMarket: ', answer);
-			answer = await SportsAMM.canExerciseMaturedMarket(awayTeamNotLoseMarket.address);
-			console.log('Can exercise options awayTeamNotLoseMarket: ', answer);
-			answer = await SportsAMM.canExerciseMaturedMarket(noDrawMarket.address);
-			console.log('Can exercise options noDrawMarket: ', answer);
-
 			let balancesHomeTeamNotLose = await homeTeamNotLoseMarket.balancesOf(first);
 
 			let payoutOnCancelationHomeTeamNotLose =
@@ -1256,247 +1249,6 @@ contract('SportsAMM DoubleChance', (accounts) => {
 				'get odds for all active markets',
 				await SportPositionalMarketData.getOddsForAllActiveMarkets()
 			);
-		});
-
-		it('Detailed test from creation to resolution - double chance', async () => {
-			let odds = [];
-			odds[0] = await SportsAMM.obtainOdds(homeTeamNotLoseMarket.address, 0);
-			odds[1] = await SportsAMM.obtainOdds(awayTeamNotLoseMarket.address, 1);
-			odds[2] = await SportsAMM.obtainOdds(noDrawMarket.address, 2);
-			console.log(
-				'Game odds: homeTeamNotLoseAnswer=',
-				fromUnit(odds[0]),
-				', awayTeamNotLoseMarket=',
-				fromUnit(odds[1]),
-				', noDrawMarket=',
-				fromUnit(odds[2])
-			);
-
-			let user1_position = 0; // 1X
-			let user1_USDamount = 100;
-			let user2_position = 0; // 2X
-			let user2_USDamount = 100;
-			let user3_position = 0; // 12
-			let user3_USDamount = 100;
-
-			let availableToBuy = await SportsAMM.availableToBuyFromAMM(
-				homeTeamNotLoseMarket.address,
-				user1_position
-			);
-			console.log(
-				'Available to buy for user 1 position homeTeamNotLoseMarket: ',
-				fromUnit(availableToBuy)
-			);
-			availableToBuy = await SportsAMM.availableToBuyFromAMM(
-				awayTeamNotLoseMarket.address,
-				user2_position
-			);
-			console.log(
-				'Available to buy for user 2 position awayTeamNotLoseMarket: ',
-				fromUnit(availableToBuy)
-			);
-			availableToBuy = await SportsAMM.availableToBuyFromAMM(noDrawMarket.address, user3_position);
-			console.log('Available to buy for user 3 position noDrawMarket: ', fromUnit(availableToBuy));
-
-			let additionalSlippage = toUnit(0.05);
-			console.log('Additional slipage: ', fromUnit(additionalSlippage));
-			let buyFromAmmQuote_1 = await SportsAMM.buyFromAmmQuote(
-				homeTeamNotLoseMarket.address,
-				user1_position,
-				toUnit(user1_USDamount)
-			);
-			console.log(
-				'User 1 buy quote homeTeamNotLoseMarket for ',
-				user1_USDamount,
-				': ',
-				fromUnit(buyFromAmmQuote_1)
-			);
-
-			let buyFromAmmQuote_2 = await SportsAMM.buyFromAmmQuote(
-				awayTeamNotLoseMarket.address,
-				user2_position,
-				toUnit(user2_USDamount)
-			);
-			console.log(
-				'User 2 buy quote awayTeamNotLoseMarket for ',
-				user2_USDamount,
-				': ',
-				fromUnit(buyFromAmmQuote_2)
-			);
-
-			let buyFromAmmQuote_3 = await SportsAMM.buyFromAmmQuote(
-				noDrawMarket.address,
-				user3_position,
-				toUnit(user3_USDamount)
-			);
-
-			console.log(
-				'User 3 buy quote noDrawMarket for ',
-				user3_USDamount,
-				': ',
-				fromUnit(buyFromAmmQuote_3)
-			);
-
-			let balance = await Thales.balanceOf(first);
-			console.log('USD balance of user 1: ', fromUnit(balance));
-			balance = await Thales.balanceOf(second);
-			console.log('USD balance of user 2: ', fromUnit(balance));
-			balance = await Thales.balanceOf(third);
-			console.log('USD balance of user 3: ', fromUnit(balance));
-			balance = await Thales.balanceOf(SportsAMM.address);
-			console.log('USD balance of AMM: ', fromUnit(balance));
-			console.log('User 1, User 2, User 3 buying ....');
-
-			answer = await SportsAMM.buyFromAMM(
-				homeTeamNotLoseMarket.address,
-				user1_position,
-				toUnit(user1_USDamount),
-				buyFromAmmQuote_1,
-				additionalSlippage,
-				{ from: first }
-			);
-			answer = await SportsAMM.buyFromAMM(
-				awayTeamNotLoseMarket.address,
-				user2_position,
-				toUnit(user2_USDamount),
-				buyFromAmmQuote_2,
-				additionalSlippage,
-				{ from: second }
-			);
-
-			answer = await SportsAMM.buyFromAMM(
-				noDrawMarket.address,
-				user3_position,
-				toUnit(user3_USDamount),
-				buyFromAmmQuote_3,
-				additionalSlippage,
-				{ from: third }
-			);
-
-			let options = await homeTeamNotLoseMarket.balancesOf(first);
-			console.log(
-				'User 1 options bought: ',
-				fromUnit(options[0]),
-				fromUnit(options[1]),
-				fromUnit(options[2])
-			);
-			options = await awayTeamNotLoseMarket.balancesOf(second);
-			console.log(
-				'User 2 options bought: ',
-				fromUnit(options[0]),
-				fromUnit(options[1]),
-				fromUnit(options[2])
-			);
-
-			options = await noDrawMarket.balancesOf(third);
-			console.log(
-				'User 3 options bought: ',
-				fromUnit(options[0]),
-				fromUnit(options[1]),
-				fromUnit(options[2])
-			);
-			console.log('-------------------------------------------');
-			balance = await Thales.balanceOf(first);
-			console.log('USD balance of user 1: ', fromUnit(balance));
-			balance = await Thales.balanceOf(second);
-			console.log('USD balance of user 2: ', fromUnit(balance));
-			balance = await Thales.balanceOf(third);
-			console.log('USD balance of user 3: ', fromUnit(balance));
-			balance = await Thales.balanceOf(SportsAMM.address);
-			console.log('USD balance of AMM: ', fromUnit(balance));
-
-			console.log('-------------------------------------------');
-			console.log('-------------- RESOLVE GAME ---------------');
-			console.log('-------------------------------------------');
-			await fastForward(await currentTime());
-			assert.equal(true, await homeTeamNotLoseMarket.canResolve());
-			assert.equal(true, await awayTeamNotLoseMarket.canResolve());
-			assert.equal(true, await noDrawMarket.canResolve());
-
-			const tx_2 = await TherundownConsumerDeployed.fulfillGamesResolved(
-				reqIdResolve,
-				gamesResolved,
-				sportId_4,
-				{ from: wrapper }
-			);
-			let gameR = await TherundownConsumerDeployed.gameResolved(gameFootballid1);
-			let marketAdd = await TherundownConsumerDeployed.marketPerGameId(gameFootballid1);
-			const tx_resolve = await TherundownConsumerDeployed.resolveMarketManually(
-				marketAdd,
-				2,
-				1,
-				2,
-				{ from: owner }
-			);
-			answer = await deployedMarket.result();
-			let game_results = ['Cancelled', 'Home', 'Away', 'Draw'];
-			console.log('Game result: ', game_results[parseInt(answer.toString())], ' wins');
-			marketAdd = await TherundownConsumerDeployed.marketPerGameId(gameid1);
-
-			options = await homeTeamNotLoseMarket.balancesOf(first);
-			console.log('User 1 options to excercise: ', fromUnit(options[user1_position]));
-			options = await awayTeamNotLoseMarket.balancesOf(second);
-			console.log('User 2 options to excercise: ', fromUnit(options[user2_position]));
-			options = await noDrawMarket.balancesOf(third);
-			console.log('User 3 options to excercise: ', fromUnit(options[user3_position]));
-
-			answer = await Thales.balanceOf(first);
-			let initial_balance_1 = answer;
-			balance = await Thales.balanceOf(first);
-			console.log('USD balance of user 1 before excercising: ', fromUnit(balance));
-			balance = await Thales.balanceOf(second);
-			let initial_balance_2 = balance;
-			console.log('USD balance of user 2 before excercising: ', fromUnit(balance));
-			balance = await Thales.balanceOf(third);
-			let initial_balance_3 = balance;
-			console.log('USD balance of user 3 before excercising: ', fromUnit(balance));
-			console.log('----------- EXCERCISING OPTIONS -----------');
-
-			options = await deployedMarket.balancesOf(homeTeamNotLoseMarket.address);
-			console.log(
-				'homeTeamNotLoseMarket options: ',
-				fromUnit(options[0]),
-				fromUnit(options[1]),
-				fromUnit(options[2])
-			);
-			options = await deployedMarket.balancesOf(awayTeamNotLoseMarket.address);
-			console.log(
-				'awayTeamNotLoseMarket options: ',
-				fromUnit(options[0]),
-				fromUnit(options[1]),
-				fromUnit(options[2])
-			);
-			options = await deployedMarket.balancesOf(noDrawMarket.address);
-			console.log(
-				'noDrawMarket options: ',
-				fromUnit(options[0]),
-				fromUnit(options[1]),
-				fromUnit(options[2])
-			);
-
-			await homeTeamNotLoseMarket.exerciseOptions({ from: first });
-			await awayTeamNotLoseMarket.exerciseOptions({ from: second });
-			await noDrawMarket.exerciseOptions({ from: third });
-
-			options = await homeTeamNotLoseMarket.balancesOf(first);
-			console.log('User 1 options after excercise: ', fromUnit(options[user1_position]));
-			options = await awayTeamNotLoseMarket.balancesOf(second);
-			console.log('User 2 options after excercise: ', fromUnit(options[user2_position]));
-			options = await noDrawMarket.balancesOf(second);
-			console.log('User 3 options after excercise: ', fromUnit(options[user2_position]));
-
-			balance = await Thales.balanceOf(first);
-			console.log('USD balance of user 1 after excercising: ', fromUnit(balance));
-			let cost = balance.sub(initial_balance_1);
-			console.log('User 1 gained after excercising: ', fromUnit(cost));
-			balance = await Thales.balanceOf(second);
-			console.log('USD balance of user 2 after excercising: ', fromUnit(balance));
-			cost = balance.sub(initial_balance_2);
-			console.log('User 2 gained after excercising: ', fromUnit(cost));
-			balance = await Thales.balanceOf(third);
-			console.log('USD balance of user 3 after excercising: ', fromUnit(balance));
-			cost = balance.sub(initial_balance_3);
-			console.log('User 3 gained after excercising: ', fromUnit(cost));
 		});
 
 		it('Buy from SportsAMM double chance - check spent on game', async () => {
